Crate pallet_block_rewards

BlockRewards Pallet

The BlockRewards pallet provides functionality for distributing rewards to different accounts with different currencies. The distribution happens when an session (a constant time interval) finalizes. Users cannot stake manually as their collator membership is syncronized via a provider. Thus, when new collators join, they will automatically be staked and vice-versa when collators leave, they are unstaked.

The BlockRewards pallet provides functions for:

Claiming the reward given for a staked currency. The reward will be the native network’s token.

Admin methods to configure the reward amount for collators and an optional beneficiary.

Re-exports

pub use weights::WeightInfo;

pub use pallet::*;

Modules

migrations

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

weights

Autogenerated weights for pallet_block_rewards

Structs

CollatorChanges

SessionChanges

Type that contains the pending update.

SessionData

Type that contains the associated data of an session