Trait pallet_block_rewards::pallet::Config

pub trait Config: Config {

    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type AdminOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type Balance: Balance + MaxEncodedLen + FixedPointOperand + Into<<<Self as Config>::Currency as CurrencyT<Self::AccountId>>::Balance> + MaybeSerializeDeserialize;
    type ExistentialDeposit: Get<Self::Balance>;
    type Weight: Parameter + MaxEncodedLen + EnsureAdd + Unsigned + FixedPointOperand + Default;
    type Rewards: GroupRewards<Balance = Self::Balance, GroupId = u32> + AccountRewards<Self::AccountId, Balance = Self::Balance, CurrencyId = <Self as Config>::CurrencyId> + CurrencyGroupChange<GroupId = u32, CurrencyId = <Self as Config>::CurrencyId>;
    type Currency: Mutate<Self::AccountId, AssetId = <Self as Config>::CurrencyId, Balance = Self::Balance> + CurrencyT<Self::AccountId>;
    type CurrencyId: Parameter + Member + Copy + MaybeSerializeDeserialize + Ord + TypeInfo + MaxEncodedLen;
    type StakeCurrencyId: Get<<Self as Config>::CurrencyId>;
    type StakeAmount: Get<<Self as Config>::Balance>;
    type StakeGroupId: Get<u32>;
    type MaxChangesPerSession: Get<u32> + TypeInfo + Debug + Clone + PartialEq;
    type MaxCollators: Get<u32> + TypeInfo + Debug + Clone + PartialEq;
    type Beneficiary: OnUnbalanced<<<Self as Config>::Currency as CurrencyT<<Self as Config>::AccountId>>::NegativeImbalance>;
    type AuthorityId: Member + Parameter + RuntimeAppPublic + MaybeSerializeDeserialize + MaxEncodedLen;
    type WeightInfo: WeightInfo;

}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

type AdminOrigin: EnsureOrigin<Self::RuntimeOrigin>

Required origin for admin purposes for configuring groups and currencies.

type Balance: Balance + MaxEncodedLen + FixedPointOperand + Into<<<Self as Config>::Currency as CurrencyT<Self::AccountId>>::Balance> + MaybeSerializeDeserialize

Type used to handle balances.

type ExistentialDeposit: Get<Self::Balance>

type Weight: Parameter + MaxEncodedLen + EnsureAdd + Unsigned + FixedPointOperand + Default

Type used to handle group weights.

type Rewards: GroupRewards<Balance = Self::Balance, GroupId = u32> + AccountRewards<Self::AccountId, Balance = Self::Balance, CurrencyId = <Self as Config>::CurrencyId> + CurrencyGroupChange<GroupId = u32, CurrencyId = <Self as Config>::CurrencyId>

The reward system used.

type Currency: Mutate<Self::AccountId, AssetId = <Self as Config>::CurrencyId, Balance = Self::Balance> + CurrencyT<Self::AccountId>

The type used to handle currency minting and burning for collators.

type CurrencyId: Parameter + Member + Copy + MaybeSerializeDeserialize + Ord + TypeInfo + MaxEncodedLen

The currency type of the artificial block rewards currency.

type StakeCurrencyId: Get<<Self as Config>::CurrencyId>

The identifier of the artificial block rewards currency which is minted and burned for collators.

type StakeAmount: Get<<Self as Config>::Balance>

The amount of the artificial block rewards currency which is minted and burned for collators.

type StakeGroupId: Get<u32>

The identifier of the collator group.

type MaxChangesPerSession: Get<u32> + TypeInfo + Debug + Clone + PartialEq

Max number of changes of the same type enqueued to apply in the next session. Max calls to Pallet::set_collator_reward() or to Pallet::set_total_reward() with the same id.

type MaxCollators: Get<u32> + TypeInfo + Debug + Clone + PartialEq

type Beneficiary: OnUnbalanced<<<Self as Config>::Currency as CurrencyT<<Self as Config>::AccountId>>::NegativeImbalance>

Target of receiving non-collator-rewards. NOTE: If set to none, collators are the only group receiving rewards.

type AuthorityId: Member + Parameter + RuntimeAppPublic + MaybeSerializeDeserialize + MaxEncodedLen

The identifier type for an authority.

type WeightInfo: WeightInfo

Information of runtime weightsk